move output names to _helpers
